+++ libraries/javascript/perspective-client/src/hooks/index.ts
@@ -2,10 +2,6 @@ import { ComponentDomEvents } from './useDomEvents'
 import { ComponentLifecycleEvents } from './useLifecycleEvents'
 
 export * from './useComponentEvents'
-export * from './useDeepCompareCallback'
-export * from './useDeepCompareEffect'
-export * from './useDeepCompareLayoutEffect'
-export * from './useDeepCompareMemo'
 export * from './useDomEvents'
 export * from './useFirstMountState'
 export * from './useLifecycleEvents'
