+++ libraries/javascript/perspective-client/src/index.ts
@@ -1,4 +1,5 @@
 export * from './components'
+export * from './globals'
 export * from './hooks'
 export * from './scripting'
 export * from './transforms'
