+++ libraries/javascript/perspective-client/src/utils/emitStyles/index.ts
