+++ libraries/javascript/perspective-client/src/utils/formatStyleNames/formatStyleNames.test.ts
