+++ libraries/javascript/perspective-client/src/utils/formatStyleNames/index.ts
