+++ libraries/javascript/perspective-client/src/utils/index.ts
@@ -1,3 +1,5 @@
+export { default as emitStyles } from './emitStyles'
+export { default as formatStyleNames } from './formatStyleNames'
 export { default as getClientStore } from './getClientStore'
 export { default as getDesignerStore } from './getDesignerStore'
 export { default as waitForClientStore } from './waitForClientStore'
