+++ libraries/javascript/utils/src/index.ts
@@ -1,11 +1,2 @@
 export * from './hooks'
-export { default as isAsyncFunction } from './isAsyncFunction'
-export { default as isCSSVar } from './isCSSVar'
-export { default as isFunction } from './isFunction'
-export { default as readCSSVar } from './readCSSVar'
-export {
-  default as toUserScript,
-  type UserScript,
-  type UserScriptParams,
-} from './toUserScript'
-export { default as transformProps, type PropTransform } from './transformProps'
+export * from './utils'
