+++ libraries/javascript/utils/src/utils/isAsyncFunction/index.ts
