+++ libraries/javascript/utils/src/utils/isAsyncFunction/isAsyncFunction.test.ts
@@ -1,7 +1,5 @@
 import { it, expect, describe } from 'vitest'
-
-import isFunction from '../../../../../libraries/javascript/utils/src/isAsyncFunction/isAsyncFunction'
-import isAsyncFunction from '../../../../../libraries/javascript/utils/src/isAsyncFunction/isAsyncFunction'
+import isAsyncFunction from '../isAsyncFunction'
 
 describe('isAsyncFunction', () => {
   it('false if no arrow function', async () => {
@@ -13,21 +11,21 @@ describe('isAsyncFunction', () => {
   })
 
   it('true if arrow function with extra spaces', async () => {
-    expect(isFunction('  async  () => arrow function here          ')).toBe(
-      true
-    )
+    expect(
+      isAsyncFunction('  async  () => arrow function here          ')
+    ).toBe(true)
   })
 
   it('true if arrow function with new lines', async () => {
     expect(
-      isFunction(`  async  () => arrow 
+      isAsyncFunction(`  async  () => arrow 
         function here          `)
     ).toBe(true)
   })
 
   it('true if arrow function parameters', async () => {
     expect(
-      isFunction(
+      isAsyncFunction(
         '    async(parameter1, parameter2) => arrow function here          '
       )
     ).toBe(true)
