+++ libraries/javascript/utils/src/utils/isAsyncFunction/isAsyncFunction.ts
