+++ libraries/javascript/utils/src/utils/isCSSVar/isCSSVar.ts
