+++ libraries/javascript/utils/src/utils/isFunction/index.ts
