+++ libraries/javascript/utils/src/utils/isFunction/isFunction.ts
