+++ libraries/javascript/utils/src/utils/uniqueClasses/index.ts
