+++ modules/charts/web/src/components/apexcharts/ApexChartsComponent.tsx
@@ -18,10 +18,9 @@ import {
   getScriptTransform,
   JavaScriptRunEvent,
   useComponentEvents,
-  useDeepCompareMemo,
   useRefLifecycleEvents,
 } from '@embr-js/perspective-client'
-import { transformProps } from '@embr-js/utils'
+import { transformProps, useDeepCompareMemo } from '@embr-js/utils'
 import { ApexChartProps, Chart } from './react/ApexCharts'
 import ApexCharts, { ApexOptions } from 'apexcharts'
