+++ modules/periscope/web/src/extensions/toast/toast.tsx
@@ -1,14 +1,15 @@
-import { toast, ToastContainer } from 'react-toastify'
-import { createRoot } from 'react-dom/client'
-import { getEmbrGlobals } from '@embr-js/perspective-client/src/globals'
 import { merge } from 'lodash'
+import { observer } from 'mobx-react-lite'
+import React from 'react'
+import { createRoot } from 'react-dom/client'
+import { toast, ToastContainer } from 'react-toastify'
+
+import { getEmbrGlobals } from '@embr-js/perspective-client'
+import { useMediaQuery } from '@embr-js/utils'
+import { ClientStore } from '@inductiveautomation/perspective-client'
+import { DockOffset } from '@inductiveautomation/perspective-client/build/dist/typedefs/stores/MountStore'
 
 import './toast.css'
-import React from 'react'
-import { ClientStore } from '@inductiveautomation/perspective-client'
-import { observer } from 'mobx-react-lite'
-import { DockOffset } from '@inductiveautomation/perspective-client/build/dist/typedefs/stores/MountStore'
-import { useMediaQuery } from '@embr-js/utils'
 
 type CenterToastContainerProps = {
   dockOffset: DockOffset
