+++ modules/periscope/web/src/util/index.ts
@@ -1,6 +1,1 @@
 export * from './JoinableView'
-export { default as emitStyles } from './emitStyles'
-export { default as formatStyleNames } from './formatStyleNames'
-export { default as mergeStyles } from './mergeStyles'
-export { default as resolve } from './resolve'
-export { default as uniqueClasses } from './uniqueClasses'
