+++ modules/periscope/web/vite.config.ts
@@ -40,6 +40,7 @@ export default defineConfig(({ mode }) => ({
     },
   },
   test: {
+    passWithNoTests: true,
     fileParallelism: mode !== 'benchmark',
     globals: true,
     environment: 'node',
